code/requirements.txt
matplotlib==3.8.0
missingno==0.5.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.0.1
Pillow==10.3.0
plotly==5.22.0
pyecharts==2.0.5
seaborn==0.13.2
streamlit==1.33.0
streamlit_echarts==0.4.0
wordcloud==1.9.3
